# web/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Partner Api",
        "version": "1.0.0"
    },
    "host": "api5.audi.dev.agence-one.net",
    "basePath": "/partner/dev/",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/service/types": {
            "get": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "Get Aftersales Services types",
                "responses": {
                    "200": {
                        "description": "Aftersales Services types",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/service/{id}": {
            "get": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "Get an AftersalesService by id",
                "operationId": "getAftersalesServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "aftersalesServiceId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target AftersalesService",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/AftersalesServiceComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "update AftersalesService by id",
                "operationId": "putAftersalesServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "aftersalesServiceId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "aftersalesService",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AftersalesService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated AftersalesService",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/AftersalesServiceComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "updating error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "remove AftersalesService by id",
                "operationId": "removeAftersalesServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "aftersalesServiceId",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete acknowledge",
                        "schema": {
                            "$ref": "#/definitions/Success"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "patch AftersalesService by id",
                "operationId": "patchAftersalesServiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "aftersalesServiceId",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "aftersalesService",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AftersalesService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the updated AftersalesService",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/AftersalesServiceComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "updating error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/service/": {
            "get": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "get AftersalesService collection",
                "operationId": "cgetAftersalesService",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "max items per page",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Target AftersalesServices",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/AftersalesServiceComplete"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "AftersalesService"
                ],
                "summary": "create AftersalesService",
                "operationId": "createAftersalesService",
                "parameters": [
                    {
                        "name": "aftersales_service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AftersalesService"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "the created AftersalesService",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/AftersalesServiceComplete"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/types": {
            "get": {
                "tags": [
                    "Partner"
                ],
                "summary": "Get partner type",
                "responses": {
                    "200": {
                        "description": "Types contract partner",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Context"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "Context": {
            "properties": {
                "context": {
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "Success": {
            "properties": {
                "data": {
                    "properties": {
                        "message": {
                            "type": "string"
                        }
                    },
                    "type": "object"
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
        "Error": {
            "required": [
                "errors"
            ],
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "properties": {
                            "message": {
                                "type": "string"
                            },
                            "code": {
                                "type": "integer"
                            },
                            "type": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "type": "object",
            "allOf": [
                {
                    "$ref": "#/definitions/Context"
                }
            ]
        },
        "IdableEntity": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "TimestampableEntity": {
            "required": [
                "createdAt",
                "updatedAt"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "updatedAt": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "AftersalesService": {
            "required": [
                "type",
                "name"
            ],
            "properties": {
                "type": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "AftersalesServiceComplete": {
            "allOf": [
                {
                    "$ref": "#/definitions/IdableEntity"
                },
                {
                    "$ref": "#/definitions/AftersalesService"
                },
                {
                    "$ref": "#/definitions/TimestampableEntity"
                }
            ]
        }
    },
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}